move BaseLoggedInPage out of cfme package
