+++ cfme/automate/explorer/__init__.py
@@ -6,8 +6,8 @@ from widgetastic.widget import View
 from widgetastic_manageiq import Accordion, ManageIQTree
 from widgetastic_patternfly import Dropdown, FlashMessages
 
-from cfme import BaseLoggedInPage
 from cfme.base import Server
+from cfme.base.login import BaseLoggedInPage
 from cfme.base.ui import automate_menu_name
 from utils.appliance.implementations.ui import navigator, CFMENavigateStep
