+++ cfme/control/explorer/__init__.py
@@ -4,8 +4,8 @@ from widgetastic.widget import View
 from widgetastic_manageiq import ManageIQTree
 from widgetastic_patternfly import Accordion, Dropdown
 
-from cfme import BaseLoggedInPage
 from cfme.base import Server
+from cfme.base.login import BaseLoggedInPage
 from utils.appliance.implementations.ui import navigator, CFMENavigateStep
