+++ cfme/control/import_export.py
@@ -5,7 +5,7 @@ from navmazing import NavigateToSibling
 from widgetastic.widget import Select, ClickableMixin
 from widgetastic_patternfly import BootstrapSelect, Button, Input
 
-from cfme import BaseLoggedInPage
+from cfme.base.login import BaseLoggedInPage
 from cfme.base.ui import Server
