+++ cfme/services/catalogs/__init__.py
@@ -3,7 +3,7 @@ from widgetastic.widget import View
 from widgetastic_manageiq import Accordion, ManageIQTree
 from widgetastic_patternfly import Dropdown
 
-from cfme import BaseLoggedInPage
+from cfme.base.login import BaseLoggedInPage
 from cfme.base import Server
 from utils.appliance.implementations.ui import navigator, CFMENavigateStep
