+++ cfme/services/workloads.py
@@ -5,7 +5,7 @@ from navmazing import NavigateToAttribute, NavigateToSibling
 from widgetastic.widget import View, Text
 from widgetastic_manageiq import Accordion, ManageIQTree, Search
 
-from cfme import BaseLoggedInPage
+from cfme.base.login import BaseLoggedInPage
 from utils.appliance import Navigatable
 from cfme.base import Server
 from utils.appliance.implementations.ui import navigator, CFMENavigateStep
